lead-qualifier/requirements.txt
# Pipeline API + run_pipeline dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM (DeepSeek via OpenAI client — primary; Anthropic optional for Claude fallback)
openai>=1.0.0
anthropic>=0.40.0

# Scraping
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
